integration-tests/requirements.txt
git+https://github.com/ptoscano/pytest-client-tools@main
pyyaml
pytest-subtests
